Move WindowFrame class from org.apache.pinot.query.runtime.operator.window.aggregate to org.apache.pinot.query.runtime.operator.window
