+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/operator/window/WindowFrame.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.query.runtime.operator.window.aggregate;
+package org.apache.pinot.query.runtime.operator.window;
 
 import org.apache.pinot.query.planner.plannode.WindowNode;
