+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/operator/window/range/DenseRankWindowFunction.java
@@ -23,7 +23,7 @@ import java.util.List;
 import org.apache.calcite.rel.RelFieldCollation;
 import org.apache.pinot.common.utils.DataSchema;
 import org.apache.pinot.query.planner.logical.RexExpression;
-import org.apache.pinot.query.runtime.operator.window.aggregate.WindowFrame;
+import org.apache.pinot.query.runtime.operator.window.WindowFrame;
 
 
 public class DenseRankWindowFunction extends RankBasedWindowFunction {
